requirements.txt
blinker==1.8.2
cachelib==0.13.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Mail==0.10.0
Flask-Session==0.8.0
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
packaging==24.1
protobuf==5.27.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
rsa==4.9
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
